+++ src/app/Playground/VetleSamples/VetlePage.xaml.cs
@@ -6,6 +6,7 @@ using DIPS.Mobile.UI.Components.ListItems;
 using DIPS.Mobile.UI.Components.Lists;
 using DIPS.Mobile.UI.Extensions;
 using DIPS.Mobile.UI.Resources.Icons;
+using Microsoft.Maui.Controls.PlatformConfiguration;
 using Playground.HåvardSamples;
 using CollectionView = DIPS.Mobile.UI.Components.Lists.CollectionView;
 using Shell = DIPS.Mobile.UI.Components.Shell.Shell;
@@ -17,7 +18,7 @@ public partial class VetlePage
     public VetlePage()
     {
         InitializeComponent();
-        TestCommand = new Command(SwitchRoot);
+        TestCommand = new Command(SwitchRoot); 
 
         _ = FuckThisShitUp();
     }
@@ -72,6 +73,7 @@ public partial class VetlePage
     private string m_test1 = "Test123";
     private double? m_originalHeight;
     private bool m_ignoreScrollEvent;
+    private bool m_firstTime;
 
     public bool TestBool
     {
@@ -100,9 +102,9 @@ public partial class VetlePage
         var test = Platform.GetCurrentUIViewController();
 #endif
 
-        
-        
-        /*_ = Navigation.PushModalAsync(new NavigationPage(new VetleTestPage1()));*/
+        /*if(!m_firstTime)
+            */
+        m_firstTime = true;
         /*Button.IsVisible = true;*/
 
         /*
@@ -227,6 +229,8 @@ public partial class VetlePage
     {
         /*SearchBar.HasCancelButton = !SearchBar.HasCancelButton;*/
         /*SearchBar.IsVisible = !SearchBar.IsVisible;*/
+        _ = Navigation.PushModalAsync(new NavigationPage(new VetleTestPage1()));
     }
 
+    
 }