+++ src/library/DIPS.Mobile.UI/API/Builder/Android/AppHostBuilderExtensions.cs
@@ -1,10 +1,17 @@
+using Android.Views;
+using AndroidX.AppCompat.Graphics.Drawable;
 using DIPS.Mobile.UI.API.Camera.Preview;
 using DIPS.Mobile.UI.API.Camera.Preview.Android.Slider;
 using DIPS.Mobile.UI.API.Library;
 using DIPS.Mobile.UI.Components.Chips;
-using DIPS.Mobile.UI.Components.Tabs;
+using Microsoft.Maui.Controls.PlatformConfiguration;
+using Microsoft.Maui.Handlers;
 using Microsoft.Maui.LifecycleEvents;
+using Microsoft.Maui.Platform;
+using DIPS.Mobile.UI.Components.Tabs;
 using Tab = Microsoft.Maui.Controls.Tab;
+using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
+using Shell = DIPS.Mobile.UI.Components.Shell.Shell;
 
 namespace DIPS.Mobile.UI.API.Builder;
 
@@ -19,6 +26,28 @@ public static partial class AppHostBuilderExtensions
             .AddHandler<CameraZoomSlider, CameraZoomSliderHandler>()
             .AddHandler<CameraPreview, CameraPreviewHandler>()
             .AddHandler<Tab, TabHandler>();
+        
+        ToolbarHandler.Mapper.AppendToMapping<Toolbar, IToolbarHandler>(nameof(Toolbar.ToolbarItems), (h, t) =>
+        {
+            // TODO: Workaround: .NET MAUI does not set the color on Text toolbar items on Shell pages, so we need to set it manually.
+            // TODO: Workaround: .NET MAUI does not set the color on Text and icon toolbar items on Modal pages, so we need to set it manually.
+            for (var i = 0; i < h.PlatformView.Menu?.Size(); i++)
+            {
+                var item = h.PlatformView.Menu.GetItem(i);
+                var span = new Android.Text.SpannableString(item?.TitleFormatted);
+                span.SetSpan(new Android.Text.Style.ForegroundColorSpan(Colors.GetColor(Shell.ForegroundColorName).ToPlatform()), 0, span.Length(), 0);
+                item?.SetTitle(span);
+            }
+        });
+        
+        ToolbarHandler.Mapper.AppendToMapping<Toolbar, IToolbarHandler>(nameof(Toolbar.BackButtonVisible), (h, t) =>
+        {
+            // TODO: Modal Workaround: .NET MAUI does not inherit the color from the Shell, so we need to set it manually.
+            if (h.PlatformView.NavigationIcon is DrawerArrowDrawable icon)
+            {
+                icon.Color = Colors.GetColor(Shell.ForegroundColorName).ToPlatform();
+            }
+        });
     }
 
     static partial void ConfigurePlatformLifecycleEvents(ILifecycleBuilder events)