+++ src/library/DIPS.Mobile.UI/Components/ListItems/Options/Subtitle/SubtitleOptions.Properties.cs
@@ -86,6 +86,6 @@ public partial class SubtitleOptions
         nameof(TextColor),
         typeof(Color),
         typeof(ListItem),
-        defaultValue:Colors.GetColor(ColorName.color_text_subtle_small));
+        defaultValue:Colors.GetColor(ColorName.color_text_subtle));
    
 }