+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## [46.0.0] 
+- Set new colors.
+- Updated colors.
+- Obsoleted old colors.
+- [iOS] Can now use Large Title. (However, it's buggy)
+- [Android] Removed workaround to set status bar when pushing modal
+- [Android] Removed workaround for Label
+- Upgraded to .NET MAUI 9.0.80.
+
+## [45.11.0] 
+- Resources was updated from DIPS.Mobile.DesignTokens
+
+## [45.10.0] 
+- Resources was updated from DIPS.Mobile.DesignTokens
+
+## [45.9.7]
+- [MultiItemsPicker] Can now tap outside the `BottomSheet` to close.
+- [BottomSheet] `SearchBar` has now keyboard `Done` button as default.
+
 ## [45.9.6]
 - [BottomSheet][Android] Focusing any `SearchBar` in `BottomSheet` will now expand the `BottomSheet`.